# docs/specs/openapi.yaml
components:
  schemas:
...
    TonBalanceResponse:
      type: "object"
      properties:
        balanceNano:
          type: "string"
        address:
          type: "string"
    StakeInfo:
      type: "object"
      properties:
        tier:
          type: "string"
          enum:
          - "FLEXIBLE"
          - "SILVER"
          - "GOLD"
          - "DIAMOND"
        amount:
          type: "string"
          description: "Stake amount in nano (decimal string)"
        startTime:
          type: "integer"
          format: "int64"
        unlockTime:
          type: "integer"
          format: "int64"
        lastClaimTime:
          type: "integer"
          format: "int64"
        pendingRewards:
          type: "string"
          description: "Pending rewards in nano (decimal string)"
        unlocked:
          type: "boolean"
    TierConfigDto:
      type: "object"
      properties:
        tier:
          type: "string"
          enum:
          - "FLEXIBLE"
          - "SILVER"
          - "GOLD"
          - "DIAMOND"
        lockDurationSec:
          type: "integer"
          format: "int64"
        multiplier:
          type: "number"
          format: "double"
        rewardSharePercent:
          type: "integer"
          format: "int32"
      description: "Staking lock config for one tier"
    UserStakingProfile:
      type: "object"
      properties:
        address:
          type: "string"
        highestTier:
          type: "string"
          enum:
          - "FLEXIBLE"
          - "SILVER"
          - "GOLD"
          - "DIAMOND"
        totalStakedNano:
          type: "string"
          description: "Total staked nano (decimal string)"
        votingPowerNano:
          type: "string"
          description: "Voting power nano (decimal string)"
        stakes:
          type: "array"
          items:
            $ref: "#/components/schemas/StakeInfo"
        tierConfigs:
          type: "array"
          items:
            $ref: "#/components/schemas/TierConfigDto"
        liveTierTvls:
          type: "object"
          additionalProperties:
            type: "integer"
            description: "Per-tier TVL nano strings; omitted keys were not available"
          description: "Per-tier TVL nano strings; omitted keys were not available"
      description: "Staking snapshot; omit address for catalog-only (empty stakes)"
    JettonInfoResponse:
      type: "object"
      properties:
        circulatingNano:
          type: "string"
        mintable:
          type: "boolean"
    FeeParamsResponse:
      type: "object"
      properties:
        burnBps:
          type: "integer"
          format: "int32"
        stakingBps:
          type: "integer"
          format: "int32"
        treasuryBps:
          type: "integer"
          format: "int32"
    ExcludedTransferResponse:
      type: "object"
      properties:
        excluded:
          type: "boolean"
...
info:
  title: "Burned Chats API"
  description: "REST surface for Burned Chats Telegram Mini App"
  version: "0.4.4"
openapi: "3.0.1"
paths:
...
  /api/wallet/burn-balance:
    get:
      tags:
      - "wallet-controller"
      operationId: "burnBalance"
      parameters:
      - name: "address"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/wallet/excluded-transfer:
    get:
      tags:
      - "wallet-controller"
      summary: "Whether sender or recipient is on the BURN master excluded list"
      operationId: "excludedTransfer"
      parameters:
      - name: "sender"
        in: "query"
        description: "Sender TON wallet (friendly or raw)"
        required: false
        schema:
          type: "string"
      - name: "recipient"
        in: "query"
        description: "Recipient TON wallet; omit to check sender only"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Excluded-transfer flag"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExcludedTransferResponse"
        "400":
          description: "Missing or invalid TON address"
          content:
            '*/*':
              schema:
                type: "object"
        "502":
          description: "Master RPC exhausted"
          content:
            '*/*':
              schema:
                type: "object"
  /api/wallet/fee-params:
    get:
      tags:
      - "wallet-controller"
      summary: "Effective BURN fee split in basis points"
      operationId: "feeParams"
      responses:
        "200":
          description: "Fee split"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FeeParamsResponse"
        "502":
          description: "Master RPC exhausted"
          content:
            '*/*':
              schema:
                type: "object"
  /api/wallet/jetton-info:
    get:
      tags:
      - "wallet-controller"
      summary: "BURN jetton circulating supply (nano decimal string) and mintable\
        \ flag"
      operationId: "jettonInfo"
      responses:
        "200":
          description: "Master supply snapshot"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JettonInfoResponse"
        "502":
          description: "Master RPC exhausted"
          content:
            '*/*':
              schema:
                type: "object"
  /api/wallet/jetton-wallet:
    get:
      tags:
      - "wallet-controller"
      operationId: "jettonWallet"
      parameters:
      - name: "address"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/wallet/staking-profile:
    get:
      tags:
      - "wallet-controller"
      summary: "Staking snapshot (positions + catalog); omit address for catalog-only"
      operationId: "stakingProfile"
      parameters:
      - name: "address"
        in: "query"
        description: "TON wallet; omit for catalog-only"
        required: false
        schema:
          type: "string"
      - name: "fresh"
        in: "query"
        description: "1 = bust user profile + computed get-keys (rate-limited 1/15s)"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "Snapshot or catalog-only"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserStakingProfile"
        "400":
          description: "Invalid TON address"
          content:
            '*/*':
              schema:
                type: "object"
        "502":
          description: "User RPC or catalog RPC exhausted"
          content:
            '*/*':
              schema:
                type: "object"
  /api/wallet/ton-balance:
    get:
      tags:
      - "wallet-controller"
      summary: "Native TON balance in nano (decimal string)"
      operationId: "tonBalance"
      parameters:
      - name: "address"
        in: "query"
        description: "TON wallet (friendly or raw)"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "TON balance"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TonBalanceResponse"